Avoid compile "Error:(37, 46) No matching Shape found..." (still other compile errors). Change name to parkingSLot.

+++ app/models/persistence/SlickTables.scala
@@ -32,7 +32,7 @@ object SlickTables extends HasDatabaseConfig[JdbcProfile] {
 
     def freeFrom = column[String]("free_from")
 
-    def userId = column[Double]("user_id")
+    def userId = column[Long]("user_id")
 
     def * = (id, lat, lon, freeFrom, userId) <> (ParkingSlot.tupled, ParkingSlot.unapply _)
   }
